Move tracking into conditions #66
